Document Intelligence/doc_intel/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
pymupdf
python-docx
requests
google-genai
python-dotenv
openai
pillow
pytesseract
python-pptx
openpyxl
pandas
beautifulsoup4
langchain
langgraph
langchain_openai
langchain_community
langchain_core
langchain_pymupdf4llm
tf-keras
